pyproject.toml
[project]
name = "dhis2-era5land-simple"
version = "0.1.0"
description = "Simple script to import ERA5-Land climate data into DHIS2"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dhis2-client",
    "dhis2eo",
    "earthkit",
    "ipykernel",
    "metpy",
    "papermill",
    "python-dotenv",
]

[tool.uv.sources]
dhis2eo = { git = "https://github.com/dhis2/dhis2eo.git" }
dhis2-client = { git = "https://github.com/dhis2/dhis2-python-client.git" }

[dependency-groups]
dev = [
    "ruff>=0.8",
    "mypy>=1.0",
    "pyright>=1.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
